+++ internal/encrypt/level2.go
@@ -7,13 +7,13 @@ import (
 	// "strings"
 )
 
-type Level1 struct {
+type Level2 struct {
 	password string
 	key1 string
 	key2 string
 }
 
-func NewLevel1(password string, key1 string) *Level1 {
+func NewLevel2(password string, key1 string, key2 string) *Level2 {
 	if password == "" {
 		// TODO: Thow an error
 	}
@@ -21,5 +21,8 @@ func NewLevel1(password string, key1 string) *Level1 {
 				key1 = fmt.Sprintf("%d", rand.Intn(90000)+1000)
 
 	}
-	return &Level1{password: password, key1: key1}
+	if key2 == "" {
+				key2 = fmt.Sprintf("%d", rand.Intn(90000)+1000)
+	}
+	return &Level2{password: password, key1: key1, key2: key2}
 }
