+++ internal/encrypt/level3.go
@@ -3,17 +3,17 @@ package encrypt
 import (
 	"fmt"
 	"math/rand"
-	"strconv"
-	"strings"
+	// "strconv"
+	// "strings"
 )
 
-type Level2 struct {
+type Level3 struct {
 	password string
 	key1 string
 	key2 string
 }
 
-func NewLevel2(password string, key1 string, key2 string) *Level2 {
+func NewLevel3(password string, key1 string, key2 string) *Level3 {
 	if password == "" {
 		// TODO: Thow an error
 	}
@@ -24,5 +24,5 @@ func NewLevel2(password string, key1 string, key2 string) *Level2 {
 	if key2 == "" {
 				key2 = fmt.Sprintf("%d", rand.Intn(90000)+1000)
 	}
-	return &Level2{password: password, key1: key1, key2: key2}
+	return &Level3{password: password, key1: key1, key2: key2}
 }
