--- internal/encryption/encrypt.go
@@ -1,1 +1,0 @@
-package encrypt
